Add support for custom box-shadow values with corrected positioning (#346)
Directly applying a box-shadow to the spinner lines would cause the shadow to be rotated along with the lines, which doesn't look natural. Previously, this was avoided by generating another element for each line, coloring it black and applying a black box-shadow, and then positioning it down a couple pixels beneath the line. However, this approach didn't allow customization of the shadow position and color, and also blocked use of inset shadows or multiple blended shadows with different positions.

This commit solves these issues by parsing a standard box-shadow value to determine its x and y offsets, and then using trigonometry to correct the shadow coordinates for each line based on its degrees of rotation.

Native CSS scaling is now used so that shadows will be scaled along with the lines as expected.

Resolves #35.